% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.bed.R
\name{read.bed}
\alias{read.bed}
\title{Read BED files}
\usage{
read.bed(bed.file, trackline = FALSE, sep = "\\t")
}
\arguments{
\item{bed.file}{path to the BED file.}

\item{trackline}{is a trackline included in the BED file passed to \code{bed.file}. If yes,
then the first line is skipped.}

\item{sep}{column delimiter.}
}
\description{
Function to read files in BED file format.
}
\examples{
bed.file <- system.file("bed_example.bed", package = "seqreadr")
 
# read BED file  
test.bed <- read.bed(bed.file)
# look at the BED file
test.bed
}
\author{
Hajk-Georg Drost
}
\references{
http://www.ensembl.org/info/website/upload/bed.html
}

