% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.proteome.R
\name{read.proteome}
\alias{read.proteome}
\title{Read the proteome of a given organism}
\usage{
read.proteome(file, format, ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the proteome.}

\item{format}{a character string specifying the file format used to store the proteome, e.g. \code{format = "fasta"} or \code{format = "fastq"}.}

\item{...}{additional arguments that are used by the \code{\link[Biostrings]{readAAStringSet}} function.}
}
\value{
A \code{\link{data.frame}} storing the \code{gene id} in the first column, the corresponding sequence as string in the second column, and the sequence length in the third column.
}
\description{
Read an organism specific proteome stored in a fasta/fastq file format.
}
\details{
This function takes a string specifying the path to the proteome file
of interest as first argument.

It is possible to read in different proteome file standards such as \emph{fasta} or \emph{fastq}.

Proteomes stored in fasta format can be downloaded from http://www.ebi.ac.uk/reference_proteomes.
}
\examples{
# reading a proteome stored in a fasta file
Ath.proteome <- read.proteome(system.file('seqs/ortho_thal_aa.fasta', package = 'seqreadr'),
                               format = "fasta")

dplyr::glimpse(Ath.proteome)
}
\author{
Hajk-Georg Drost
}

