% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.bed.R
\name{write.bed}
\alias{write.bed}
\title{Write BED files to hard drive}
\usage{
write.bed(bed.data, file.name = "file.bed")
}
\arguments{
\item{bed.data}{a \code{\link{data.frame}} fulfilling the BED file format.}

\item{file.name}{name of the BED file.}
}
\description{
Write a \code{\link{data.frame}} fulfilling the BED file format to
your hard drive.
}
\examples{

bed.file <- system.file("bed_example.bed", package = "seqreadr")
 
# read BED file  
test.bed <- read.bed(bed.file)
# write BED file to the hard drive
write.bed(test.bed)
}
\author{
Hajk-Georg Drost
}

