% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.gtf.R
\name{write.gtf}
\alias{write.gtf}
\title{Write GTF files to hard drive}
\usage{
write.gtf(gtf.data, file.name = "file.gtf")
}
\arguments{
\item{gtf.data}{a \code{\link{data.frame}} fulfilling the GTF file format.}

\item{file.name}{name of the GTF file.}
}
\description{
Write a \code{\link{data.frame}} fulfilling the GTF file format to
your hard drive.
}
\examples{

example.gtf <- system.file("example.gtf", package = "seqreadr")
 
# read GTF file  
test.gtf <- read.gtf(example.gtf)
# write GTF file to the hard drive
write.gtf(test.gtf)
}
\author{
Hajk-Georg Drost
}

